requirements.txt
# === Основные ===
fastapi==0.115.11                 # Основной фреймворк (FastAPI)
uvicorn==0.34.0                   # ASGI-сервер для запуска приложения
pydantic==2.10.6                  # Валидация и сериализация данных (схемы)

# === Работа с базой данных ===
sqlalchemy[asyncio]==2.0.39       # ORM с поддержкой asyncio
asyncpg==0.29.0                   # Асинхронный драйвер PostgreSQL
alembic==1.13.3                   # Миграции базы данных
sqladmin==0.18.0                  # Админ-панель на основе SQLAlchemy

# === Аутентификация и безопасность ===
python-jose[cryptography]==3.4.0  # JWT-токены и работа с криптографией
bcrypt==4.0.1                     # Хеширование паролей
itsdangerous==2.2.0               # Подпись и проверка токенов (ссылки, куки и др.)

# === Конфигурация и инфраструктура ===
python-dotenv==1.0.1              # Загрузка переменных окружения из .env

# === Работа с формами и файлами ===
fastapi-mail==1.5.0               # Отправка email из FastAPI

# === Тестирование и разработка ===
pytest==8.3.3                     # Основной фреймворк для тестирования
pytest-asyncio==0.24.0            # Поддержка асинхронных тестов
httpx==0.27.2                     # HTTP-клиент (тестирование и интеграции)
